edit sleep: handle automatic backup
Backup file should get updated after editing as well, fix this by moving
the logic to DataModel and then using this from the Main and from the
Sleep activities.

Related to <https://github.com/vmiklos/plees-tracker/issues/138>.
