+++ app/src/main/java/hu/vmiklos/plees_tracker/MainActivity.kt
@@ -26,7 +26,6 @@ import androidx.activity.result.contract.ActivityResultContracts.RequestMultiple
 import androidx.activity.result.contract.ActivityResultContracts.StartActivityForResult
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.ContextCompat
-import androidx.documentfile.provider.DocumentFile
 import androidx.lifecycle.ViewModelProvider
 import androidx.preference.PreferenceManager
 import androidx.recyclerview.widget.DefaultItemAnimator
@@ -263,36 +262,12 @@ class MainActivity : AppCompatActivity() {
     fun startStop(@Suppress("UNUSED_PARAMETER") v: View) {
         if (DataModel.start != null && DataModel.stop == null) {
             DataModel.stop = Calendar.getInstance().time
-            viewModel.stopSleep()
-            backupSleeps()
+            viewModel.stopSleep(applicationContext, contentResolver)
         } else {
             DataModel.start = Calendar.getInstance().time
             DataModel.stop = null
         }
         updateView()
-    }
-
-    private fun backupSleeps() {
-        val preferences = DataModel.preferences
-        val autoBackup = preferences.getBoolean("auto_backup", false)
-        val autoBackupPath = preferences.getString("auto_backup_path", "")
-        if (!autoBackup || autoBackupPath == null || autoBackupPath.isEmpty()) {
-            return
-        }
-
-        val folder = DocumentFile.fromTreeUri(applicationContext, Uri.parse(autoBackupPath))
-            ?: return
-
-        // Make sure that we don't create "backup (1).csv", etc.
-        val oldBackup = folder.findFile("backup.csv")
-        if (oldBackup != null && oldBackup.exists()) {
-            oldBackup.delete()
-        }
-
-        val backup = folder.createFile("text/csv", "backup.csv") ?: return
-        viewModel.exportDataToFile(
-            applicationContext, contentResolver, backup.uri, showToast = false
-        )
     }
 
     private fun exportFileData() {
